Refactor search vs list code to clean up the main Products component.
